requirements.txt
# Core dependencies
google-genai>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0

# FastAPI backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

